+++ internal/discover/lookup_integration_test.go
@@ -21,20 +21,17 @@ var testConvJID = jid.MustParse("sam@conversations.im")
 var lookupTests = [...]struct {
 	resolver *net.Resolver
 	service  string
-	addr     net.Addr
+	addr     jid.JID
 	addrs    []*net.SRV
 	err      error
 }{
-	0: {},
+	0: {
+		err: discover.ErrInvalidService,
+	},
 	1: {
 		service: "xmpp-client",
 		addr:    jid.MustParse("me@example.net"),
-		addrs: []*net.SRV{
-			{
-				Target: "example.net",
-				Port:   5222,
-			},
-		},
+		addrs:   discover.FallbackRecords("xmpp-client", "example.net"),
 	},
 	2: {
 		service: "xmpp-client",
@@ -44,19 +41,25 @@ var lookupTests = [...]struct {
 				Target:   "xmpp.conversations.im.",
 				Port:     5222,
 				Priority: 5,
-				Weight:   1,
+				Weight:   0,
+			},
+			{
+				Target:   "xmpps.conversations.im.",
+				Port:     80,
+				Priority: 10,
+				Weight:   0,
 			},
 		},
 	},
 	3: {
 		service: "xmpp-server",
-		addr:    &testConvJID,
+		addr:    testConvJID,
 		addrs: []*net.SRV{
 			{
 				Target:   "xmpp.conversations.im.",
 				Port:     5269,
 				Priority: 5,
-				Weight:   1,
+				Weight:   0,
 			},
 		},
 	},
@@ -74,14 +77,14 @@ var lookupTests = [...]struct {
 	},
 }
 
-func TestLookupService(t *testing.T) {
+func TestIntegrationLookupService(t *testing.T) {
 	for i, tc := range lookupTests {
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			addrs, err := discover.LookupService(context.Background(), tc.resolver, tc.service, "tcp", tc.addr)
+			addrs, err := discover.LookupService(context.Background(), tc.resolver, tc.service, tc.addr)
 			switch dnsErr := err.(type) {
 			case nil:
 				if err != tc.err {
-					t.Errorf("Got unexpected error: want=%q, got=%q", tc.err, err)
+					t.Errorf("got unexpected error: want=%q, got=%q", tc.err, err)
 				}
 			case *net.DNSError:
 				var errStr string
@@ -89,19 +92,22 @@ func TestLookupService(t *testing.T) {
 					errStr = tc.err.Error()
 				}
 				if dnsErr.Err != errStr {
-					t.Errorf("Got unexpected error: want=%q, got=%q", errStr, dnsErr.Error())
+					t.Errorf("got unexpected error: want=%q, got=%q", errStr, dnsErr.Error())
 				}
 			default:
 				if err != tc.err {
-					t.Errorf("Got unexpected error: want=%q, got=%q", tc.err, err)
+					t.Errorf("got unexpected error: want=%q, got=%q", tc.err, err)
 				}
 			}
 			if len(tc.addrs) != len(addrs) {
-				t.Fatalf("Unexpected addrs: want=%d, got=%d", len(tc.addrs), len(addrs))
+				for _, addr := range addrs {
+					t.Logf("got addr: %+v", addr)
+				}
+				t.Fatalf("unexpected addrs: want=%d, got=%d", len(tc.addrs), len(addrs))
 			}
 			for i, addr := range tc.addrs {
 				if *addr != *addrs[i] {
-					t.Fatalf("Unexpected addr at %d: want=%v, got=%v", i, *addr, *addrs[i])
+					t.Fatalf("unexpected addr at %d: want=%v, got=%v", i, *addr, *addrs[i])
 				}
 			}
 		})